Factor out `embedding`
