+++ async-openai/src/types/mod.rs
@@ -6,6 +6,7 @@ mod assistant_impls;
 mod audio;
 mod common;
 mod edit;
+mod embedding;
 mod file;
 mod fine_tune;
 mod fine_tuning;
@@ -23,6 +24,7 @@ pub use assistant_file::*;
 pub use audio::*;
 pub use common::*;
 pub use edit::*;
+pub use embedding::*;
 pub use file::*;
 pub use fine_tune::*;
 pub use fine_tuning::*;
